requirements.txt
groq
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv
fastapi
uvicorn
streamlit
torch
transformers
langchain
langchain-groq